refactoring of delegate and dlite code to use same helper functions. Needs a round of testing
