+++ command/command.go
@@ -9,8 +9,8 @@ import (
 	"os"
 
 	"github.com/drone-runners/drone-runner-aws/command/daemon"
-	"github.com/drone-runners/drone-runner-aws/command/delegate"
-	"github.com/drone-runners/drone-runner-aws/command/dlite"
+	"github.com/drone-runners/drone-runner-aws/command/harness/delegate"
+	"github.com/drone-runners/drone-runner-aws/command/harness/dlite"
 	"github.com/drone-runners/drone-runner-aws/command/setup"
 
 	"gopkg.in/alecthomas/kingpin.v2"
