+++ command/harness/dlite/init.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"net/http"
 
-	"github.com/harness/lite-engine/api"
+	"github.com/drone-runners/drone-runner-aws/command/harness"
 	"github.com/wings-software/dlite/client"
 	"github.com/wings-software/dlite/logger"
 )
@@ -14,23 +14,9 @@ type VmInitTask struct {
 	c *dliteCommand
 }
 
-type SetupVmRequest struct {
-	ID               string            `json:"id"` // stage runtime ID
-	PoolID           string            `json:"pool_id"`
-	Tags             map[string]string `json:"tags"`
-	CorrelationID    string            `json:"correlation_id"`
-	LogKey           string            `json:"log_key"`
-	api.SetupRequest `json:"setup_request"`
-}
-
-type SetupVmResponse struct {
-	IPAddress  string `json:"ip_address"`
-	InstanceID string `json:"instance_id"`
-}
-
 type VmInitRequest struct {
-	SetupVmRequest SetupVmRequest     `json:"setup_vm_request"`
-	Services       []ExecuteVmRequest `json:"services"`
+	SetupVmRequest harness.SetupVmRequest     `json:"setup_vm_request"`
+	Services       []harness.ExecuteVmRequest `json:"services"`
 }
 
 func (t *VmInitTask) ServeHTTP(w http.ResponseWriter, r *http.Request) {
@@ -57,7 +43,7 @@ func (t *VmInitTask) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	// fmt.Printf("init request: %+v\n", req)
 
 	// Make the setup call
-	setupResp, err := t.c.handleSetup(ctx, &req.SetupVmRequest)
+	setupResp, err := harness.HandleSetup(ctx, &req.SetupVmRequest, t.c.stageOwnerStore, t.c.env, t.c.poolManager)
 	if err != nil {
 		logger.WriteJSON(w, failedResponse(err.Error()), 500)
 		return
@@ -69,7 +55,7 @@ func (t *VmInitTask) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	for _, s := range req.Services {
 		s.IPAddress = setupResp.IPAddress
 		status := VmServiceStatus{ID: s.ID, Name: s.Name, Image: s.Image, LogKey: s.LogKey, Status: Running}
-		resp, err := t.c.handleStep(ctx, &s)
+		resp, err := harness.HandleStep(ctx, &s, t.c.env, t.c.poolManager)
 		if err != nil {
 			status.Status = Error
 			status.ErrorMessage = err.Error()
